Refactoring of common funcs (#4)

+++ cortado/rtas/recycle_bin_process.py
@@ -8,10 +8,13 @@
 # ATT&CK: T1158
 # Description: Executes mock malware from the "C:\Recycler\" and "C:\$RECYCLE.BIN\" subdirectories.
 
+import logging
 import time
 from pathlib import Path
 
-from . import _common, register_code_rta, OSType, RuleMetadata
+from . import OSType, RuleMetadata, _common, register_code_rta
+
+log = logging.getLogger(__name__)
 
 
 RECYCLE_PATHS = ["C:\\$Recycle.Bin", "C:\\Recycler"]
@@ -27,34 +30,37 @@ TARGET_APP_EXE = "bin/myapp.exe"
         RuleMetadata(id="cff92c41-2225-4763-b4ce-6f71e5bda5e6", name="Execution from Unusual Directory - Command Line")
     ],
     techniques=["T1036", "T1059"],
-    ancillary_files=[TARGET_APP_EXE, _common.CMD_PATH],
+    ancillary_files=[TARGET_APP_EXE],
 )
 def main():
-    _common.log("Execute files from the Recycle Bin")
+    log.info("Execute files from the Recycle Bin")
     target_dir = None
     for recycle_path in RECYCLE_PATHS:
         if Path(recycle_path).exists():
             target_dir = _common.find_writeable_directory(recycle_path)
             if target_dir:
                 break
-
     else:
-        _common.log("Could not find a writeable directory in the recycle bin")
-        exit(1)
+        log.info("Could not find a writeable directory in the recycle bin")
+        raise _common.ExecutionError("Could not find a writeable directory in the recycle bin")
 
-    commands = [
-        [TARGET_APP_EXE],
-        [_common.CMD_PATH, "/c", "echo hello world"],
-    ]
+    if not target_dir:
+        raise _common.ExecutionError("No writable directories in `RECYCLE_PATHS`")
 
-    _common.log("Running commands from recycle bin in %s" % target_dir)
-    for command in commands:  # type: list[str]
-        source_path = command[0]
-        arguments = command[1:]
+    log.info("Running commands from recycle bin in %s" % target_dir)
+    target_path = Path(target_dir) / "recycled_process.exe"
 
-        target_path = Path(target_dir) / "recycled_process.exe"
-        _common.copy_file(source_path, target_path)
-        arguments.insert(0, target_path)
-        _common.execute(arguments)
-        time.sleep(0.5)
-        _common.remove_file(target_path)
+    target_app_exe = _common.get_resource_path(TARGET_APP_EXE)
+    _common.copy_file(target_app_exe, target_path)
+    _ = _common.execute_command([target_path])
+
+    time.sleep(0.5)
+    _common.remove_file(target_path)
+
+    cmd_path = _common.get_cmd_path()
+    _common.copy_file(cmd_path, target_path)
+    _ = _common.execute_command(
+        [target_path, "/c", "echo hello world"],
+    )
+    time.sleep(0.5)
+    _common.remove_file(target_path)
